TMDM-12549 SE code is based on EE permissions which leads to inability to create… (#635)
TMDM-12549 SE code is based on EE permissions which leads to inability to create or modify any record
