+++ org.talend.mdm.core/pom.xml
@@ -69,6 +69,8 @@
                             <excludes>
                                 <exclude>com/amalto/core/delegator/impl/DefaultItemCtrlDelegator.class</exclude>
                                 <exclude>com/amalto/core/delegator/impl/DefaultValidationDelegator.class</exclude>
+                                <exclude>com/amalto/core/delegator/impl/DefaultSecurityCheckDelegator.class</exclude>
+                                <exclude>com/amalto/core/delegator/impl/DefaultMenuUtilDelegator.class</exclude>
                                 <exclude>com/amalto/core/delegator/impl/DefaultXtentisWSDelegator.class</exclude>
                                 <exclude>com/amalto/core/delegator/impl/DefaultLocalUserDelegator.class</exclude>
                                 <exclude>com/amalto/core/save/generator/AutoIncrementGenerator.class</exclude>
@@ -85,6 +87,8 @@
                             <includes>
                                 <include>com/amalto/core/delegator/impl/DefaultItemCtrlDelegator.class</include>
                                 <include>com/amalto/core/delegator/impl/DefaultValidationDelegator.class</include>
+                                <include>com/amalto/core/delegator/impl/DefaultSecurityCheckDelegator.class</include>
+                                <include>com/amalto/core/delegator/impl/DefaultMenuUtilDelegator.class</include>
                                 <include>com/amalto/core/delegator/impl/DefaultXtentisWSDelegator.class</include>
                                 <include>com/amalto/core/delegator/impl/DefaultLocalUserDelegator.class</include>
                                 <include>com/amalto/core/save/generator/AutoIncrementGenerator.class</include>
