+++ org.talend.mdm.core/resources/com/amalto/core/initdb/data/PROVISIONING.User.user
@@ -1,13 +1,33 @@
 <ii>
-    <c>PROVISIONING</c>	<n>User</n>	<dmn>PROVISIONING</dmn>	<dmr/>	<sp/>	<i>user</i>	<t>1257918109062</t>	<p>
+    <c>PROVISIONING</c>
+    <n>User</n>
+    <dmn>PROVISIONING</dmn>
+    <dmr />
+    <sp />
+    <i>user</i>
+    <t>1257918109062</t>
+    <p>
         <User>
-            <id>user</id>    <username>user</username>    <password>ee11cbb19052e40b07aac0ca060c23ee</password>    <givenname>Default</givenname>    <familyname>User</familyname>    <phonenumber/>    <company>Company</company>    <signature/>    <realemail>user@company.com</realemail>    <fakeemail/>    <viewrealemail>no</viewrealemail>    <registrationdate>1245811028031</registrationdate>    <lastvisitdate>0</lastvisitdate>    <lastsynctime>0</lastsynctime>    <enabled>yes</enabled>    <homepage>Home</homepage>    <universe/>    <roles>
-                <role>System_Interactive</role> 
-</roles>    
-           <properties>
-
-            </properties>    
-<applications>    </applications>
-</User>
-</p>
+            <id>user</id>
+            <username>user</username>
+            <password>ee11cbb19052e40b07aac0ca060c23ee</password>
+            <givenname>Default</givenname>
+            <familyname>User</familyname>
+            <phonenumber />
+            <company>Company</company>
+            <signature />
+            <realemail>user@company.com</realemail>
+            <fakeemail />
+            <viewrealemail>no</viewrealemail>
+            <registrationdate>1245811028031</registrationdate>
+            <lastvisitdate>0</lastvisitdate>
+            <lastsynctime>0</lastsynctime>
+            <enabled>yes</enabled>
+            <homepage>Home</homepage>
+            <universe />
+            <roles />
+            <properties />
+            <applications />
+        </User>
+    </p>
 </ii>
