+++ org.talend.mdm.core/src/com/amalto/core/delegator/BaseSecurityCheck.java
@@ -1,0 +1,19 @@
+/*
+ * Copyright (C) 2006-2018 Talend Inc. - www.talend.com
+ * 
+ * This source code is available under agreement available at
+ * %InstallDIR%\features\org.talend.rcp.branding.%PRODUCTNAME%\%PRODUCTNAME%license.txt
+ * 
+ * You should have received a copy of the agreement along with this program; if not, write to Talend SA 9 rue Pages
+ * 92150 Suresnes, France
+ */
+package com.amalto.core.delegator;
+
+import com.amalto.core.save.DocumentSaverContext;
+import com.amalto.core.save.SaverSession;
+import com.amalto.core.util.VetoException;
+
+public abstract class BaseSecurityCheck {
+    public void vetoableSave(SaverSession session, DocumentSaverContext context) throws VetoException {
+    }
+}
