+++ org.talend.mdm.core/src/com/amalto/core/delegator/BeanDelegatorContainer.java
@@ -16,6 +16,10 @@ public class BeanDelegatorContainer {
     private Map<String, Object> delegatorInstancePool;
 
     private static final String LOCAL_USER = "LocalUser"; //$NON-NLS-1$
+    
+    private static final String SECURITY_CHECK = "SecurityCheck"; //$NON-NLS-1$
+
+    private static final String MENU_UTIL = "MenuUtil"; //$NON-NLS-1$
 
     private static final String VALIDATION = "Validation"; //$NON-NLS-1$
 
@@ -50,6 +54,18 @@ public class BeanDelegatorContainer {
     public ILocalUser getLocalUserDelegator() {
         synchronized (delegatorInstancePool) {
             return (ILocalUser) delegatorInstancePool.get(LOCAL_USER);
+        }
+    }
+
+    public BaseSecurityCheck getSecurityCheckDelegator() {
+        synchronized (delegatorInstancePool) {
+            return (BaseSecurityCheck) delegatorInstancePool.get(SECURITY_CHECK);
+        }
+    }
+
+    public BaseMenu getMenuUtilDelegator() {
+        synchronized (delegatorInstancePool) {
+            return (BaseMenu) delegatorInstancePool.get(MENU_UTIL);
         }
     }
