+++ org.talend.mdm.core/src/com/amalto/core/server/security/ws/WSAuthorizationInterceptor.java
@@ -1,0 +1,19 @@
+/*
+ * Copyright (C) 2006-2018 Talend Inc. - www.talend.com
+ * 
+ * This source code is available under agreement available at
+ * %InstallDIR%\features\org.talend.rcp.branding.%PRODUCTNAME%\%PRODUCTNAME%license.txt
+ * 
+ * You should have received a copy of the agreement along with this program; if not, write to Talend SA 9 rue Pages
+ * 92150 Suresnes, France
+ */
+package com.amalto.core.server.security.ws;
+
+import org.apache.cxf.interceptor.security.SecureAnnotationsInterceptor;
+
+public class WSAuthorizationInterceptor extends SecureAnnotationsInterceptor {
+
+    public WSAuthorizationInterceptor() {
+        setAnnotationClassName(WebServiceRoles.class.getName());
+    }
+}
