+++ org.talend.mdm.core/test/com/amalto/core/metadata/compare/schema8_2.xsd
@@ -33,8 +33,8 @@
                 <xsd:element maxOccurs="1" minOccurs="0" name="universe" nillable="false" type="xsd:string"/>
                 <xsd:element maxOccurs="1" minOccurs="0" name="roles" nillable="false">
                     <xsd:complexType mixed="false">
-                        <xsd:sequence maxOccurs="1" minOccurs="1">
-                            <xsd:element maxOccurs="unbounded" minOccurs="1" name="role" nillable="false" type="xsd:string"/>
+                        <xsd:sequence maxOccurs="1" minOccurs="0">
+                            <xsd:element maxOccurs="unbounded" minOccurs="0" name="role" nillable="false" type="xsd:string"/>
                         </xsd:sequence>
                     </xsd:complexType>
                 </xsd:element>
