+++ org.talend.mdm.webapp.general/src/main/java/org/talend/mdm/webapp/general/server/actions/GeneralAction.java
@@ -45,6 +45,7 @@ import org.w3c.dom.NodeList;
 import com.amalto.core.delegator.ILocalUser;
 import com.amalto.core.server.ServerAccess.ServerAccessInfo;
 import com.amalto.core.util.LocalUser;
+import com.amalto.core.util.Menu;
 import com.amalto.core.util.Messages;
 import com.amalto.core.util.MessagesFactory;
 import com.amalto.core.util.SessionContextHolder;
@@ -62,7 +63,6 @@ import com.amalto.core.webservice.WSRegexDataClusterPKs;
 import com.amalto.core.webservice.WSRegexDataModelPKs;
 import com.amalto.core.webservice.XtentisPort;
 import com.amalto.webapp.core.bean.Configuration;
-import com.amalto.webapp.core.util.Menu;
 import com.amalto.webapp.core.util.SystemLocale;
 import com.amalto.webapp.core.util.SystemLocaleFactory;
 import com.amalto.webapp.core.util.SystemLocaleInitializable;
