+++ org.talend.mdm.webapp.general/src/main/java/org/talend/mdm/webapp/general/server/util/Utils.java
@@ -35,6 +35,7 @@ import org.xml.sax.SAXException;
 import com.amalto.commons.core.utils.XMLUtils;
 import com.amalto.core.delegator.ILocalUser;
 import com.amalto.core.util.LocalUser;
+import com.amalto.core.util.Menu;
 import com.amalto.core.util.Messages;
 import com.amalto.core.util.MessagesFactory;
 import com.amalto.core.webservice.WSCount;
@@ -46,7 +47,6 @@ import com.amalto.core.webservice.WSPutItem;
 import com.amalto.core.webservice.WSString;
 import com.amalto.core.webservice.WSWhereItem;
 import com.amalto.webapp.core.bean.Configuration;
-import com.amalto.webapp.core.util.Menu;
 import com.amalto.webapp.core.util.SystemLocale;
 import com.amalto.webapp.core.util.SystemLocaleFactory;
 import com.amalto.webapp.core.util.Util;
