+++ org.talend.mdm.webapp.welcomeportal/src/main/java/org/talend/mdm/webapp/welcomeportal/server/actions/WelcomePortalAction.java
@@ -47,7 +47,7 @@ import com.amalto.core.webservice.WSTransformerV2;
 import com.amalto.core.webservice.WSTransformerV2PK;
 import com.amalto.core.webservice.WSTypedContent;
 import com.amalto.webapp.core.bean.Configuration;
-import com.amalto.webapp.core.util.Menu;
+import com.amalto.core.util.Menu;
 import com.amalto.webapp.core.util.Util;
 import com.amalto.webapp.core.util.Webapp;
 
@@ -100,7 +100,7 @@ public class WelcomePortalAction implements WelcomePortalService {
      */
     private boolean isHiddenMenu(String menu) throws ServiceException {
         try {
-            TreeMap<String, Menu> subMenus = Menu.getRootMenu().getSubMenus();
+            Map<String, Menu> subMenus = Menu.getRootMenu().getSubMenus();
             for (String key : subMenus.keySet()) {
                 Menu subMenu = subMenus.get(key);
